#24 move k8s data struct out to dedicated go file
