+++ k8sData.go
@@ -1,0 +1,131 @@
+package main
+
+import (
+	"time"
+	"net"
+)
+
+//Data type representing the json schema of https://[master]/api/v1/pods
+type Data struct {
+	APIVersion string `json:"apiVersion"`
+	Items      []struct {
+		Metadata struct {
+			Annotations struct {
+				KubernetesIoCreatedBy string `json:"kubernetes.io/created-by"`
+			} `json:"annotations"`
+			CreationTimestamp time.Time `json:"creationTimestamp"`
+			GenerateName      string    `json:"generateName"`
+			Labels            struct {
+				App             string `json:"app"`
+				PodTemplateHash int64  `json:"pod-template-hash,string"`
+			} `json:"labels"`
+			Name            string `json:"name"`
+			Namespace       string `json:"namespace"`
+			OwnerReferences []struct {
+				APIVersion         string `json:"apiVersion"`
+				BlockOwnerDeletion bool   `json:"blockOwnerDeletion"`
+				Controller         bool   `json:"controller"`
+				Kind               string `json:"kind"`
+				Name               string `json:"name"`
+				UID                string `json:"uid"`
+			} `json:"ownerReferences"`
+			ResourceVersion int64  `json:"resourceVersion,string"`
+			SelfLink        string `json:"selfLink"`
+			UID             string `json:"uid"`
+		} `json:"metadata"`
+		Spec struct {
+			Containers []struct {
+				Image           string `json:"image"`
+				ImagePullPolicy string `json:"imagePullPolicy"`
+				Name            string `json:"name"`
+				Ports           []struct {
+					ContainerPort int64  `json:"containerPort"`
+					Protocol      string `json:"protocol"`
+				} `json:"ports"`
+				ReadinessProbe struct {
+					FailureThreshold int64 `json:"failureThreshold"`
+					HTTPGet          struct {
+						Path   string `json:"path"`
+						Port   int64  `json:"port"`
+						Scheme string `json:"scheme"`
+					} `json:"httpGet"`
+					PeriodSeconds    int64 `json:"periodSeconds"`
+					SuccessThreshold int64 `json:"successThreshold"`
+					TimeoutSeconds   int64 `json:"timeoutSeconds"`
+				} `json:"readinessProbe"`
+				Resources struct {
+					Limits struct {
+						CPU    string  `json:"string"`
+						Memory string `json:"memory"`
+					} `json:"limits"`
+					Requests struct {
+						CPU    string `json:"cpu"`
+						Memory string `json:"memory"`
+					} `json:"requests"`
+				} `json:"resources"`
+				TerminationMessagePath   string `json:"terminationMessagePath"`
+				TerminationMessagePolicy string `json:"terminationMessagePolicy"`
+				VolumeMounts             []struct {
+					MountPath string `json:"mountPath"`
+					Name      string `json:"name"`
+				} `json:"volumeMounts"`
+			} `json:"containers"`
+			DNSPolicy       string `json:"dnsPolicy"`
+			NodeName        string `json:"nodeName"`
+			RestartPolicy   string `json:"restartPolicy"`
+			SchedulerName   string `json:"schedulerName"`
+			SecurityContext struct {
+			} `json:"securityContext"`
+			ServiceAccount                string `json:"serviceAccount"`
+			ServiceAccountName            string `json:"serviceAccountName"`
+			TerminationGracePeriodSeconds int64  `json:"terminationGracePeriodSeconds"`
+			Tolerations                   []struct {
+				Effect            string `json:"effect"`
+				Key               string `json:"key"`
+				Operator          string `json:"operator"`
+				TolerationSeconds int64  `json:"tolerationSeconds"`
+			} `json:"tolerations"`
+			Volumes []struct {
+				GcePersistentDisk struct {
+					FsType    string `json:"fsType"`
+					Partition int64  `json:"partition"`
+					PdName    string `json:"pdName"`
+				} `json:"gcePersistentDisk"`
+				Name string `json:"name"`
+			} `json:"volumes"`
+		} `json:"spec"`
+		Status struct {
+			Conditions []struct {
+				LastProbeTime      interface{} `json:"lastProbeTime"`
+				LastTransitionTime time.Time   `json:"lastTransitionTime"`
+				Status             string        `json:"string"`
+				Type               string      `json:"type"`
+			} `json:"conditions"`
+			ContainerStatuses []struct {
+				ContainerID string `json:"containerID"`
+				Image       string `json:"image"`
+				ImageID     string `json:"imageID"`
+				LastState   struct {
+				} `json:"lastState"`
+				Name         string `json:"name"`
+				Ready        bool   `json:"ready"`
+				RestartCount int64  `json:"restartCount"`
+				State        struct {
+					Running struct {
+						StartedAt time.Time `json:"startedAt"`
+					} `json:"running"`
+				} `json:"state"`
+			} `json:"containerStatuses"`
+			HostIP    net.IP    `json:"hostIP"`
+			Phase     string    `json:"phase"`
+			PodIP     net.IP    `json:"podIP"`
+			QosClass  string    `json:"qosClass"`
+			StartTime time.Time `json:"startTime"`
+		} `json:"status"`
+	} `json:"items"`
+	Kind     string `json:"kind"`
+	Metadata struct {
+		ResourceVersion int64  `json:"resourceVersion,string"`
+		SelfLink        string `json:"selfLink"`
+	} `json:"metadata"`
+}
